restructure diagnostics into separate file, context passed by ref so reported diagnostics dont get forgotten
